+++ legislice/__init__.py
@@ -1,3 +1,5 @@
 from legislice.enactments import Enactment
+from legislice.citations import Citation
+from legislice.download import Client
 
 __version__ = "0.4.1"