+++ legislice/__init__.py
@@ -1,3 +1,5 @@
+"""Utility for downloading and comparing the text of statutes and constitutional provisions."""
+
 from legislice.enactments import Enactment
 from legislice.citations import Citation
 from legislice.download import Client