+++ tests/test_load_json.py
@@ -1,9 +1,4 @@
-import datetime
-
 import pytest
-
-from legislice.download import LegisliceDateError, LegislicePathError
-from legislice.enactments import TextQuoteSelector
 
 
 class TestUpdateEnactments: