+++ changelog.md
@@ -6,6 +6,8 @@ dev
 
 - add EnactmentGroup class
 - drop Python 3.7 support
+- import Citation and Client at top level of library
+- Client.fetch_cross_reference no longer will ignore "date" param
 
 0.4.1 (2020-12-31)
 ------------------